widen integer literals automatically

+++ testdata/pascal.go
@@ -3,7 +3,7 @@ package main
 // Pascal's triangle via the recursive binomial coefficient.
 func choose(n, k byte) uint16 {
 	if k == 0 || k == n {
-		return uint16(1)
+		return 1
 	}
 	return choose(n-1, k-1) + choose(n-1, k)
 }
